msm: Synchronize CPU frequency on thread migration
Sometimes when foreground threads, crucial for UI, migrate
from a core at higher frequency to the one running at a lower
frequency, it could badly affect performance for cases like
scrolling.
This change tries to raise the frequency of the destination core
to the same value as the source core, if it is not already running
at the same or higher frequency.  In order for the ondemand
governor to have a fair idea of load in the next sample, it also
reschedules the next sample one sampling interval from the
frequency transition after which the ondemand governor again
takes over.

Signed-off-by: Rohit Gupta <rohgup@codeaurora.org>
Signed-off-by: Mekala Natarajan <mekalan@codeaurora.org>

Conflicts:
	drivers/cpufreq/cpufreq_ondemand.c

Change-Id: I28e681f1016dd8f6208cdbb5e6627cbb72cb1d96

Conflicts:
	drivers/cpufreq/cpufreq_ondemand.c